Extract powersupply linux code
